[INI] CommandLineApp main class
