+++ src/stdafx.h
@@ -1,15 +1,1 @@
 
-#if defined _WINDOWS
-	#include <windows.h>
-#else
-#endif
-
-#define cg_use_DROPBOX_SYNC 0
-#define cg_use_FACEBOOK 0
-#define cg_use_TFicheroITV_zip 0
-#define cg_use_IDC_OpenGL2D 0
-#define cg_use_IDC_Direct2D 0
-#define cg_usar_VectorQ 0
-
-#define COMPEGPS_USAR_CODEPAGEHELPER
-#define MODULOSCOMUNES_AISLADOS
